+++ Sources/AcknowList/AcknowListViewController.swift
@@ -267,6 +267,15 @@ open class AcknowListViewController: UITableViewController {
         }
     }
 
+    open override func viewWillTransition(to size: CGSize, with coordinator: UIViewControllerTransitionCoordinator) {
+        super.viewWillTransition(to: size, with: coordinator)
+
+        coordinator.animate(alongsideTransition: nil) { _ in
+            self.configureHeaderView()
+            self.configureFooterView()
+        }
+    }
+
     // MARK: - Actions
 
     /**